refactor(version-control): extract shared diff-view core
